Set boolean field as tiny integer with a maximum display width of 1, signed and not-null. Add comments to sql file.
git-svn-id: https://svn.code.sf.net/p/web-erp/code/trunk@7041 09516a42-71f9-43be-811f-9b821cf0d246

+++ sql/mysql/upgrade4.11-4.12.sql
@@ -126,9 +126,14 @@ ALTER TABLE `qasamples`
 ALTER TABLE `sampleresults`
   ADD CONSTRAINT `sampleresults_ibfk_1` FOREIGN KEY (`testid`) REFERENCES `qatests` (`testid`);
 
+--
+-- Modifications of stockdescriptiontranslations table for longdescription translation and translated versions control
+--
 ALTER TABLE `stockdescriptiontranslations` CHANGE `descriptiontranslation` `descriptiontranslation` VARCHAR(50) CHARACTER SET utf8 COLLATE utf8_general_ci NULL DEFAULT NULL COMMENT 'Item''s short description';
+-- Add field for the long description translation of an item:
 ALTER TABLE `stockdescriptiontranslations` ADD `longdescriptiontranslation` TEXT CHARACTER SET utf8 COLLATE utf8_general_ci NULL DEFAULT NULL COMMENT 'Item''s long description';
-ALTER TABLE `stockdescriptiontranslations` ADD  `needsrevision` INT( 1 ) NOT NULL DEFAULT  '0';
+-- Add a field to mark if a description needs revision:
+ALTER TABLE `stockdescriptiontranslations` ADD `needsrevision` TINYINT(1) NOT NULL DEFAULT '0';
 
 INSERT INTO  `scripts` (`script` ,`pagesecurity` ,`description`) VALUES ('AutomaticTranslationDescriptions.php',  '15',  'Translates via Google Translator all empty translated descriptions ');
 
@@ -138,5 +143,3 @@ INSERT INTO  `scripts` (`script` ,`pagesecurity` ,`description`) VALUES ('Revisi
 
 
 UPDATE config SET confvalue='4.12' WHERE confname='VersionNumber';
-
-
